labhub.py: add migrate_issue plugin
migrate_issue plugin that adds ability to
migrate an issue from a source repo to
target repo, both owned by the org.

Issue title, issue description and all comments
are copied with a few additional details appended
to the description and comments. Source issue is
referenced in the target issue and closed after
the migration is completed.

Closes https://github.com/coala/corobo/issues/518